+++ src/presentation/components/index.ts
@@ -1,4 +1,6 @@
 export { default as LoginHeader } from './login-header/login-header';
 export { default as Input } from './input/input';
-export { default as FormError } from './form-error/form-error';
+export { default as FormStatus } from './form-status/form-status';
 export { default as Footer } from './footer/footer';
+export { default as Logo } from './logo/logo';
+export { default as Spinner } from './spinner/spinner';
